feat: hide SMS option when the user hasn't sms enabled
