+++ src/components/Dashboard/QuickActions/index.js
@@ -1,26 +1,7 @@
 import React from 'react';
 import { useIntl } from 'react-intl';
 
-export const QUICK_ACTIONS = [
-  {
-    labelId: 'dashboard.quick_actions.make_campaign',
-    linkId: 'dashboard.quick_actions.make_campaign_url',
-  },
-  {
-    labelId: 'dashboard.quick_actions.make_contact_list',
-    linkId: 'dashboard.quick_actions.make_contact_list_url',
-  },
-  {
-    labelId: 'dashboard.quick_actions.launch_automation',
-    linkId: 'dashboard.quick_actions.launch_automation_url',
-  },
-  {
-    labelId: 'dashboard.quick_actions.send_sms',
-    linkId: 'dashboard.quick_actions.send_sms_url',
-  },
-];
-
-export const QuickActions = () => {
+export const QuickActions = ({ quickActions = [] }) => {
   const intl = useIntl();
   const _ = (id, values) => intl.formatMessage({ id: id }, values);
 
@@ -28,7 +9,7 @@ export const QuickActions = () => {
     <>
       <h2 className="dp-title-col-postcard">{_('dashboard.quick_actions.section_name')}</h2>
       <ul className="dp-quick-actions">
-        {QUICK_ACTIONS.map((qa, index) => (
+        {quickActions.map((qa, index) => (
           <li key={`qa-${index}`}>
             <a href={_(qa.linkId)} target="_blank">
               <span className="action-title">{_(qa.labelId)}</span>
