+++ src/components/Dashboard/QuickActions/index.test.js
@@ -1,14 +1,15 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import { QuickActions, QUICK_ACTIONS } from '.';
+import { QuickActions } from '.';
 import IntlProvider from '../../../i18n/DopplerIntlProvider.double-with-ids-as-values';
+import { QUICK_ACTIONS } from '../Dashboard';
 
 describe('QuickActions', () => {
   it('should render quick actions', async () => {
     // Act
     render(
       <IntlProvider>
-        <QuickActions />
+        <QuickActions quickActions={QUICK_ACTIONS} />
       </IntlProvider>,
     );
